Removed AsepriteConverter, moved to Aseprite class
